fix: Refactor question model and controller

+++ aion/web/controllers/question_controller.ex
@@ -3,6 +3,7 @@ defmodule Aion.QuestionController do
 
   alias Aion.Question
   alias Aion.Answer
+  alias Aion.Subject
 
   def index(conn, _params) do
     questions = Repo.all(Question)
@@ -10,27 +11,34 @@ defmodule Aion.QuestionController do
   end
 
   def create(conn, %{"question" => question_params, "answers" => answers_content}) do
-    questionChangeset = Question.changeset(%Question{}, question_params)
+    question_subject = Repo.get(Subject, question_params["subject"])
 
-    case Repo.insert(questionChangeset) do
-      {:ok, question} ->
-        answerChangeset = Answer.changeset(%Answer{}, %{"content" => answers_content, "belongs_to" => question})
-        case Repo.insert(answerChangeset) do
-          {:ok, answers} ->
-            conn
-              |> put_status(:created)
-              |> put_resp_header("location", question_path(conn, :show, question))
-              |> render("show.json", question: question)
-          {:error, changeset} ->
-            Repo.delete(question)
-            conn
-              |> put_status(:unprocessable_entity)
-              |> render(Aion.ChangesetView, "error.json", changeset: changeset)
-        end
-      {:error, changeset} ->
-        conn
+    if question_subject do
+      question_changeset = Question.changeset(%Question{}, Map.put(question_params, "belongs_to", question_subject))
+      case Repo.insert(question_changeset) do
+        {:ok, question} ->
+          answer_changeset = Answer.changeset(%Answer{}, %{"content" => answers_content, "belongs_to" => question})
+          case Repo.insert(answer_changeset) do
+            {:ok, answer} ->
+              conn
+                |> put_status(:created)
+                |> put_resp_header("location", question_path(conn, :show, question))
+                |> render("show.json", question: question, answer: answer)
+            {:error, error_answer_changeset} ->
+              Repo.delete(question)
+              conn
+                |> put_status(:unprocessable_entity)
+                |> render(Aion.ChangesetView, "error.json", changeset: error_answer_changeset)
+          end
+        {:error, error_question_changeset} ->
+          conn
+            |> put_status(:unprocessable_entity)
+            |> render(Aion.ChangesetView, "error.json", changeset: error_question_changeset)
+      end
+    else
+      conn
         |> put_status(:unprocessable_entity)
-        |> render(Aion.ChangesetView, "error.json", changeset: changeset)
+        |> render(Aion.ChangesetView, "error.json", info: "Question subject does not exist" )
     end
   end
